Struct africastalking_gateway::AfricasTalkingGateway

pub struct AfricasTalkingGateway { /* fields omitted */ }

Gateway struct

Methods

impl AfricasTalkingGateway

pub fn new(username: &str, api_key: &str, env: &str) -> Self

creates a gateway instance

pub fn get_user_data(&self) -> Result<Value>

Gets user data

pub fn send_message(&self, msg: SMSMessage) -> Result<Value>

Sends an SMS message read more..

pub fn fetch_messages(&self, last_received_id: i32) -> Result<Vec<SMSMessage>>

Fetches messages from Africa's Talking API read more..

The gateway will return 100 messages at a time, starting with the last received id (id of the message that you last processed). Specify 0 if this is the first call to the gateway.

pub fn create_subscription(
&self,
phone_number: &str,
short_code: &str,
keyword: &str
) -> Result<Value>

Creates a subscription read more..

pub fn delete_subscription(
&self,
phone_number: &str,
short_code: &str,
keyword: &str
) -> Result<Value>

Deletes a subscription

pub fn fetch_subscriptions(
&self,
short_code: &str,
keyword: &str,
last_received_id: i32
) -> Result<Value>

Fetches subscriptions read more..

pub fn call(&self, from: &str, to: &str) -> Result<Value>

Makes voice call. docs reference

pub fn get_queued_calls(
&self,
phone_number: &str,
queue_name: Option<&str>
) -> Result<Value>

Gets queued calls. docs reference

pub fn upload_media_file(&self, media_url: &str) -> Result<Value>

Uploads Media File. docs reference

pub fn send_airtime(&self, recipients: &Value) -> Result<Value>

Sends airtime. docs reference

recipients is a json array of the format

[
  {
    "phoneNumber":"+254711XXXYYY",
    "amount":"KES X"
  },
  {
    "phoneNumber":"+254733YYYZZZ",
    "amount":"KES Y"
  }
]

pub fn init_mobile_payment_checkout(
&self,
product_name: &str,
phone_number: &str,
currency_code: &str,
provider_channel: &str,
amount: f32,
metadata: &HashMap<&str, &str>
) -> Result<Value>

Initiates a checkout request on a subscriber's phone number. read more ..

pub fn mobile_payment_b2b_request(
&self,
product_name: &str,
provider_data: &HashMap<&str, &str>,
currency_code: &str,
amount: f32,
metadata: &HashMap<&str, &str>
) -> Result<Value>

Requests a Business-to-Business payment to a business via their provider channel. read more..

pub fn mobile_payment_b2c_request(
&self,
product_name: &str,
recipients: &Value
) -> Result<Value>

Requests a Business-to-Consumer payment to mobile subscribers phone numbers. read more..